pyproject.toml
[build-system]
requires = [
    "setuptools >= 61.0",
    "wheel",
    "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "seqnado"
authors = [
    {name = "Alastair Smith", email = "alastair.smith@ndcls.ox.ac.uk"},
    {name = "Catherine Chahrour", email = "catherine.chahrour@msdtc.ox.ac.uk"}
]
description = "Pipelines for genomics analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
dynamic = ["version"]
dependencies = [
    "click<=8.3.1",
    "cookiecutter<=2.6.0",
    "flask<=3.0.0",
    "gunicorn<=23.0.0",

    "numpy<=2.3.5",
    "pandas<=2.3.3",
    "pandera<=0.27.0",
    "pexpect<=4.9.0",
    "pulp<=3.3.0",
    "pydantic<=2.12.5",
    "pyyaml<=6.0.3",
    "rich<=14.2.0",
    "seaborn<=0.13.2",
    "snakemake-wrapper-utils<=0.8.0",
    "snakemake>=9.12.0,<=9.14.5",
    "tracknado<=0.2.4",
    "typer<=0.20.0",
    "wget<=3.2",
]

[project.optional-dependencies]
slurm = ["snakemake-executor-plugin-slurm<=2.0.3"]

[project.scripts]
seqnado = "seqnado.cli:app"
seqnado-web = "seqnado_web.cli:main"


[project.urls]
Homepage = "https://github.com/alsmith151/seqnado"

[tool.setuptools_scm]
write_to = "seqnado/_version.py"
local_scheme = "no-local-version"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests (fast, isolated)",
    "integration: Integration tests (slower, multiple components)",
    "edge_case: Edge case tests",
    "slow: Slow tests (may take significant time)",
    "requires_data: Tests that require test data files",
    "parametrize: Parametrized tests",
    "pipeline: End-to-end Snakemake pipeline tests (opt-in)",
    "snakemake: Tests that invoke Snakemake via subprocess",
    "requires_apptainer: Tests that require Apptainer/Singularity",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["seqnado"]
omit = [
    "seqnado/_version.py",
    "seqnado/workflow/scripts/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if typing.TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod",
]
